special cases, not callbacks, for money and horses.
