+++ res/core/resources/horse.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0"?>
 <resource name="horse" limited="yes">
   <item big="yes" weight="5000" score="10" capacity="7000" animal="yes">
-    <construction skill="training" minskill="1"/>
-    <function name="give" value="givehorses"/>
+    <construction skill="training" minskill="1" />
   </item>
 </resource>
