+++ res/e3a/items.xml
@@ -62,13 +62,12 @@
   </resource>
 
   <resource name="charger">
-    <item big="yes" weight="5000" score="10" capacity="7000" animal="yes">
+    <item big="yes" weight="5000" score="20" capacity="7000" animal="yes">
       <construction skill="training" minskill="4">
         <requirement type="money" quantity="200"/>
         <requirement type="iron" quantity="1"/>
         <requirement type="horse" quantity="1"/>
       </construction>
-      <function name="give" value="givehorses"/>
     </item>
   </resource>
