+++ src/give.c
@@ -19,6 +19,7 @@
 
 /* kernel includes */
 #include <kernel/ally.h>
+#include <kernel/build.h>
 #include <kernel/curse.h>
 #include <kernel/faction.h>
 #include <kernel/item.h>
@@ -147,17 +148,45 @@ int give_quota(const unit * src, const unit * dst, const item_type * type,
     return n;
 }
 
+static int
+give_horses(unit * s, unit * d, const item_type * itype, int n,
+    struct order *ord)
+{
+    if (d == NULL) {
+        region *r = s->region;
+        if (r->land) {
+            rsethorses(r, rhorses(r) + n);
+        }
+        return 0;
+    }
+    return -1;                    /* use the mechanism */
+}
+
+static int
+give_money(unit * s, unit * d, const item_type * itype, int n,
+    struct order *ord)
+{
+    if (d == NULL) {
+        region *r = s->region;
+        if (r->land) {
+            rsetmoney(r, rmoney(r) + n);
+        }
+        return 0;
+    }
+    return -1;                    /* use the mechanism */
+}
+
 int
 give_item(int want, const item_type * itype, unit * src, unit * dest,
 struct order *ord)
 {
     short error = 0;
-    int n, r;
+    int n, delta;
 
     assert(itype != NULL);
     n = get_pooled(src, item2resource(itype), GET_SLACK | GET_POOLED_SLACK, want);
     n = MIN(want, n);
-    r = n;
+    delta = n;
     if (dest && src->faction != dest->faction
         && src->faction->age < GiveRestriction()) {
         if (ord != NULL) {
@@ -178,18 +207,19 @@ struct order *ord)
     else if (itype->flags & ITF_CURSED) {
         error = 25;
     }
-    else if (itype->give == NULL || itype->give(src, dest, itype, n, ord) != 0) {
+    else {
         int use = use_pooled(src, item2resource(itype), GET_SLACK, n);
+
         if (use < n)
             use +=
             use_pooled(src, item2resource(itype), GET_POOLED_SLACK,
             n - use);
         if (dest) {
-            r = give_quota(src, dest, itype, n);
-            i_change(&dest->items, itype, r);
+            delta = give_quota(src, dest, itype, n);
+            i_change(&dest->items, itype, delta);
 #ifdef RESERVE_GIVE
 #ifdef RESERVE_DONATIONS
-            change_reservation(dest, itype, r);
+            change_reservation(dest, itype, delta);
 #else
             if (src->faction == dest->faction) {
                 change_reservation(dest, item2resource(itype), r);
@@ -199,14 +229,25 @@ struct order *ord)
 #if MUSEUM_MODULE && defined(TODO)
             /* TODO: use a trigger for the museum warden! */
             if (a_find(dest->attribs, &at_warden)) {
-                warden_add_give(src, dest, itype, r);
+                warden_add_give(src, dest, itype, delta);
             }
 #endif
             handle_event(dest->attribs, "receive", src);
         }
+        else {
+            /* return horses to the region */
+            if (itype->construction && itype->flags & ITF_ANIMAL) {
+                if (itype->construction->skill == SK_HORSE_TRAINING) {
+                    give_horses(src, dest, itype, n, ord);
+                }
+            }
+            else if (itype->rtype == get_resourcetype(R_SILVER)) {
+                give_money(src, dest, itype, n, ord);
+            }
+        }
         handle_event(src->attribs, "give", dest);
     }
-    add_give(src, dest, n, r, item2resource(itype), ord, error);
+    add_give(src, dest, n, delta, item2resource(itype), ord, error);
     if (error)
         return -1;
     return 0;
