+++ src/kernel/item.test.c
@@ -162,7 +162,6 @@ static void test_core_resources(CuTest *tc) {
     CuAssertPtrNotNull(tc, rtype = rt_find("money"));
     CuAssertPtrNotNull(tc, rtype->itype);
     CuAssertPtrNotNull(tc, rtype->uchange);
-    CuAssertPtrNotNull(tc, rtype->itype->give);
     CuAssertPtrNotNull(tc, rtype = rt_find("peasant"));
     CuAssertPtrEquals(tc, 0, rtype->itype);
     CuAssertPtrNotNull(tc, rtype = rt_find("person"));
